// src/app/api/ui/enums/signed-out-of-ep-reason-type.enum.ts
export enum SignedOutOfEpReasonTypeEnum {
  UtilitiesAndEnergyServices = 1,
  InformationSystemsMaintenance = 2,
  FixedTelephoneServices = 3,
  MobileCommunicationServices = 4,
  PersonalServicesContracts = 5,
  PostalSuppliesAndServices = 6,
  SpecialAndCourierPostalServices = 7,
  InformationSecurityServices = 8,
  CostEstimateVerification = 9,
  EventParticipationServices = 10,
  NotaryServices = 11,
  BusinessTripArrangements = 12,
  WarrantyVehicleMaintenance = 13,
  EmergencyVehicleRepairs = 14,
  HeavyEquipmentTransportForRepair = 15,
  OfficialRegistryDocumentRequests = 16,
  PeriodicalPublicationsSubscription = 17,
  OfficialPublicationAdvertising = 18,
  MediaContentProductionAndPublication = 19,
  RealEstateLeasing = 20,
  GovernmentRepresentationServices = 21,
  ExtemporaneousMedicines = 22,
  NarcoticSubstancesManagement = 23,
  MedicalSupportServices = 24,
  UtilityNetworksMaintenance = 25,
  MeasurementAndLabEquipmentServices = 26,
  FireAndEmergencyResponse = 27,
  ITAndOfficeEquipmentRepair = 28,
  EmergencyApplianceRepairs = 29,
  ArchivalAndDocumentServices = 30,
  POBoxRental = 31,
  PsychiatricExaminationServices = 32,
  OfficialResidenceMaintenance = 33,
  PublicSafetyEnhancement = 34,
  UrgentSmallValueProcurement = 35,
}